Move ValueMetaJson to Core library #4645
