+++ core/src/main/java/org/apache/hop/core/row/value/ValueMetaBinary.java
@@ -21,7 +21,7 @@ import org.apache.hop.core.exception.HopValueException;
 import org.apache.hop.core.row.IValueMeta;
 
 @ValueMetaPlugin(id = "8", name = "Binary", description = "Binary", image = "images/binary.svg")
-public class ValueMetaBinary extends ValueMetaBase implements IValueMeta {
+public class ValueMetaBinary extends ValueMetaBase {
 
   public ValueMetaBinary() {
     this(null);
