+++ core/src/main/java/org/apache/hop/core/row/value/ValueMetaInteger.java
@@ -21,7 +21,7 @@ import org.apache.hop.core.exception.HopValueException;
 import org.apache.hop.core.row.IValueMeta;
 
 @ValueMetaPlugin(id = "5", name = "Integer", description = "Integer", image = "images/number.svg")
-public class ValueMetaInteger extends ValueMetaBase implements IValueMeta {
+public class ValueMetaInteger extends ValueMetaBase {
 
   public ValueMetaInteger() {
     this(null);
