+++ core/src/main/java/org/apache/hop/core/row/value/ValueMetaString.java
@@ -22,7 +22,7 @@ import org.apache.hop.core.exception.HopValueException;
 import org.apache.hop.core.row.IValueMeta;
 
 @ValueMetaPlugin(id = "2", name = "String", description = "String", image = "images/string.svg")
-public class ValueMetaString extends ValueMetaBase implements IValueMeta {
+public class ValueMetaString extends ValueMetaBase {
 
   public ValueMetaString() {
     this(null);
