--- plugins/valuetypes/json/src/assembly/assembly.xml
@@ -1,51 +1,0 @@
-<!--
-  ~ Licensed to the Apache Software Foundation (ASF) under one or more
-  ~ contributor license agreements.  See the NOTICE file distributed with
-  ~ this work for additional information regarding copyright ownership.
-  ~ The ASF licenses this file to You under the Apache License, Version 2.0
-  ~ (the "License"); you may not use this file except in compliance with
-  ~ the License.  You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  ~
-  -->
-
-<assembly xmlns="http://maven.apache.org/ASSEMBLY/2.2.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/ASSEMBLY/2.2.0 http://maven.apache.org/xsd/assembly-2.2.0.xsd">
-
-    <id>hop-valuetypes-json</id>
-    <formats>
-        <format>zip</format>
-    </formats>
-    <baseDirectory>.</baseDirectory>
-    <files>
-        <file>
-            <source>${project.basedir}/src/main/resources/version.xml</source>
-            <outputDirectory>plugins/valuetypes/json</outputDirectory>
-            <filtered>true</filtered>
-        </file>
-    </files>
-
-    <fileSets>
-        <fileSet>
-            <directory>${project.basedir}/src/main/samples</directory>
-            <outputDirectory>config/projects/samples/</outputDirectory>
-        </fileSet>
-    </fileSets>
-
-    <dependencySets>
-        <dependencySet>
-            <includes>
-                <include>org.apache.hop:hop-valuetypes-json:jar</include>
-            </includes>
-            <outputDirectory>plugins/valuetypes/json</outputDirectory>
-        </dependencySet>
-    </dependencySets>
-</assembly>
